requirements.txt
aiogram==3.0.0b7
python-dotenv==1.0.0
asyncpg==0.28.0
sqlalchemy[asyncio]==2.0.20
